==> 02_EDA_1.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

train = pd.read_parquet('data/processed/train_merged.parquet')
train.shape
train.isFraud.value_counts()

# Key Entity Columns

entity_cols = [
    col for col in train.columns
    if col.startswith('card') or
       col.startswith('addr') or
       "email" in col.lower() or
       "Device" in col or
       "id_" in col
]

entity_cols

train[entity_cols].head()

# Checking for the missing values. Only for the useful ones not all columns

missing = train[entity_cols + ['TransactionAmt', 'dt', 'isFraud']].isnull().mean().sort_values(ascending=False)
missing.head(20)
missing.to_csv('data/processed/missing_summary.csv', index=False)

# Cardinality Check
train[entity_cols].nunique().sort_values(ascending=True).head(20)

# Cardinality Check
unique_values = train[entity_cols].nunique().sort_values(ascending=False)
unique_values.to_csv('data/processed/cardinality_summary.csv', index=False)

# Correlation / Mutual Information Overview

from sklearn.feature_selection import mutual_info_classif

sample = train.sample(50000)
X = sample.select_dtypes(include=[ 'int32', 'float32']).drop('isFraud', axis=1)
y = sample['isFraud']

mi = mutual_info_classif(X.fillna(0), y)
pd.Series(mi, index=X.columns).sort_values(ascending=False).head(20)

# Extract time columns for behavioural windows

train['hour'] = train['dt'].dt.hour
train['day'] = (train['TransactionDT'] // (24*60*60)).astype(int)
train['week'] = train['dt'].dt.isocalendar().week.astype(int)

train.head(8)

train.loc[(train['card1']==12695), ['card1', 'addr2', 'DeviceInfo', 'P_emaildomain', 'isFraud']]

train.addr2.unique()

train.loc[(train['card1']==13926) & (train['addr1']==264), ['card1', 'addr1', 'DeviceInfo', 'P_emaildomain', 'isFraud']]

- PRIMARY_ENTITY = 'card1'
- SECONDARY_ENTITY = ['card1','addr1']
- DEVICE_ENTITY = 'DeviceInfo'
- EMAIL_ENTITY = 'P_emaildomain'
